refactorizacion de codigo para hacerlo testeable creacion de archivo server.js con los metodos http
